feat: gifticonId, userId에 매칭되는 기프티콘이 없는 경우 에러를 던지도록 구현

+++ src/main/java/yapp/buddycon/app/gifticon/application/service/DeleteGifticonService.java
@@ -3,18 +3,22 @@ package yapp.buddycon.app.gifticon.application.service;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import yapp.buddycon.app.gifticon.adapter.GifticonException;
 import yapp.buddycon.app.gifticon.application.port.in.DeleteGifticonUsecase;
 import yapp.buddycon.app.gifticon.application.port.out.GifticonCommandStorage;
+import yapp.buddycon.app.gifticon.application.port.out.GifticonQueryStorage;
 
 @Service
 @RequiredArgsConstructor
 @Transactional
 public class DeleteGifticonService implements DeleteGifticonUsecase {
 
+    private final GifticonQueryStorage queryStorage;
     private final GifticonCommandStorage commandStorage;
 
     @Override
     public void delete(Long userId, Long gifticonId) {
-        commandStorage.delete(userId, gifticonId);
+        if(queryStorage.existByUserIdAndGifticonId(userId, gifticonId)) commandStorage.delete(userId, gifticonId);
+        else throw new GifticonException(GifticonException.GifticonExceptionCode.GIFTICON_NOT_FOUND);
     }
 }
